+++ apps/native/app.json
@@ -1,6 +1,6 @@
 {
   "expo": {
-    "version": "0.1.28.alpha.cdc-20.6",
+    "version": "0.1.28.alpha.cdc-20.7",
     "runtimeVersion": "0.1.28",
     "orientation": "portrait",
     "icon": "./assets/images/icon.png",
